requirements.txt
numpy==1.19.2
Pillow==7.2.0
pyglet==1.5.7
SQLAlchemy==1.3.19
